Solve Problem 014 (the collatz sequence one)
Tried the simple brute-force way, and a memoization scheme. The
memoization is faster, but some of the solutions in the Project Euler
solutions thread are running in <10 ms, while my solution is around 1s.
Not sure what gains there are left to make. Maybe a larger recursion
limit so I don't have to manually manage the stack? I can't believe that
would make a 100x difference.