Solve problem 044
I do not think this is a very robust solution. The solution it finds is
the pair (P_i, P_j) such that P_i-P_j and P_i+P_j are both pentagonal,
but it minimizes P_i rather than P_i - P_j. This happens to work out.